backend/requirements.txt
django==4.0.4
django-debug-toolbar==3.4.0
Pillow==9.1.0
python-dotenv==0.20.0
dj-database-url==0.5.0
django-phonenumber-field==6.1.0
phonenumbers==8.12.47
djangorestframework==3.13.1
requests==2.27.1
geopy==2.2.0
gunicorn==20.1.0
rollbar==0.16.2
psycopg2-binary==2.9.3
